refactor: extract security sanitizer module
